+++ package.json
@@ -288,6 +288,7 @@
 	},
 	"dependencies": {
 		"adm-zip": "^0.5.14",
+		"find-process": "^1.4.7",
 		"get-port": "^7.1.0",
 		"glob": "^11.0.0",
 		"node-window-manager": "^2.2.4",
