+++ src/lib/follow_cursor.ts
@@ -1,6 +1,6 @@
 import * as vscode from 'vscode'
-import { get_config } from './util'
-import { RenpyProcess } from './process'
+import { get_config } from './config'
+import { AnyProcess } from './process'
 import { get_logger } from './logger'
 import path from 'upath'
 import { find_game_root } from './sh'
@@ -60,7 +60,7 @@ export async function sync_editor_with_renpy({
 }
 
 export async function warp_renpy_to_cursor(
-	rp: RenpyProcess,
+	rp: AnyProcess,
 	status_bar: StatusBar
 ): Promise<void> {
 	const editor = vscode.window.activeTextEditor
@@ -95,13 +95,13 @@ export class FollowCursor {
 	private status_bar: StatusBar
 	private text_editor_handle: vscode.Disposable | undefined
 
-	active_process: RenpyProcess | undefined
+	active_process: AnyProcess | undefined
 
 	constructor({ status_bar }: { status_bar: StatusBar }) {
 		this.status_bar = status_bar
 	}
 
-	async set(process: RenpyProcess) {
+	async set(process: AnyProcess) {
 		if (get_config('renpyExtensionsEnabled') !== 'Enabled') {
 			vscode.window.showErrorMessage(
 				"Follow cursor only works with Ren'Py extensions enabled.",
